prevent unintentionally overwriting existing files using foundation

+++ foundation.py
@@ -25,4 +25,8 @@ if __name__ == '__main__':
             print(f"ignore {item.name}")
         else:
             print(f"copying {item.name} to {args.dest+item.name}")
-            shutil.copyfile(item.name, args.dest+item.name)
+            if os.path.exists(args.dest+item.name):
+                print(f"adding a .donotuse to the file name so as not to overwrite existing file with same name")
+                shutil.copyfile(item.name, args.dest+item.name+".donotuse")
+            else:
+                shutil.copyfile(item.name, args.dest+item.name)
